Review by @sk1p
* prime_numba_cache() moved out of MIT code
* Extract parts of the UDF interface that are relevant for the Negotiator
* Move win_tweaks to common
* Include @sk1p's script
* Include license check in check list
