+++ .github/PULL_REQUEST_TEMPLATE
@@ -9,6 +9,7 @@
 ## Reviewer Checklist:
 
 * [ ] `/azp run libertem.libertem-data` passed
+* [ ] No import of GPL code from MIT code
 
 <!--
