+++ docs/source/releasing.rst
@@ -48,6 +48,8 @@ When planning a release, create a new issue with the following checklist:
     ## Before (using a release candidate package)
 
     * [ ] Review open issues and pull requests
+    * [ ] License review: no import of GPL code from MIT code
+          `pydeps --only "libertem" --show-deps --noshow src\libertem | python scripts\licensecheck.py`
     * [ ] Run full CI pipeline, including slow tests, on [Azure DevOps](https://dev.azure.com/LiberTEM/LiberTEM/_build?definitionId=3)
     * [ ] Run tests for related packages w/ new LiberTEM version
         * [ ] LiberTEM-live
