+++ src/libertem/cli_tweaks.py
@@ -3,5 +3,5 @@ import platform
 
 def console_tweaks():
     if platform.system() == "Windows":
-        from libertem.win_tweaks import disable_quickedit
+        from libertem.common.win_tweaks import disable_quickedit
         disable_quickedit()
