+++ src/libertem/web/dataset.py
@@ -5,14 +5,54 @@ from typing import Dict
 
 import tornado.web
 
+import numpy as np
+
 from libertem.io.dataset import load, detect, get_dataset_cls
-from libertem.common.numba import prime_numba_cache
 from .base import CORSMixin, log_message
 from libertem.common.async_utils import sync_to_async
 from .messages import Message
 from .state import SharedState
 
 log = logging.getLogger(__name__)
+
+
+def prime_numba_cache(ds):
+    dtypes = (np.float32, None)
+    for dtype in dtypes:
+        roi = np.zeros(ds.shape.nav, dtype=bool).reshape((-1,))
+        roi[max(-ds._meta.sync_offset, 0)] = True
+
+        from libertem.udf.sum import SumUDF
+        from libertem.udf.raw import PickUDF
+        from libertem.io.corrections.corrset import CorrectionSet
+        from libertem.io.dataset.base import Negotiator
+
+        # need to have at least one UDF; here we run for both sum and pick
+        # to reduce the initial latency when switching to pick mode
+        udfs = [SumUDF(), PickUDF()]
+        neg = Negotiator()
+        for udf in udfs:
+            for corr_dtype in (np.float32, None):
+                if corr_dtype is not None:
+                    corrections = CorrectionSet(dark=np.zeros(ds.shape.sig, dtype=corr_dtype))
+                else:
+                    corrections = None
+                found_first_tile = False
+                for p in ds.get_partitions():
+                    if found_first_tile:
+                        break
+                    p.set_corrections(corrections)
+                    tiling_scheme = neg.get_scheme(
+                        udfs=[udf],
+                        dataset=ds,
+                        approx_partition_shape=p.shape,
+                        read_dtype=dtype,
+                        roi=roi,
+                        corrections=corrections,
+                    )
+                    for t in p.get_tiles(tiling_scheme=tiling_scheme, roi=roi):
+                        found_first_tile = True
+                        break
 
 
 class DataSetDetailHandler(CORSMixin, tornado.web.RequestHandler):
