+++ tox.ini
@@ -5,7 +5,7 @@ envlist = flake8, py{36,37,38,39}, py{36,37,38,39}-data, benchmark, benchmark-cu
 commands=
     pytest --durations=10 --cov=libertem --cov-report=term --cov-report=html --cov-report=xml --junitxml=junit.xml {posargs:tests/}
     # win_tweaks.py depends on modules that are only available on Windows
-    pytest --doctest-modules --ignore=src/libertem/win_tweaks.py src/libertem/
+    pytest --doctest-modules --ignore=src/libertem/common/win_tweaks.py src/libertem/
 deps=
     -rtest_requirements.txt
     # Inject release candidates for testing
